dom/xhr/tests/test_sharedworker_xhr.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for SharedWorker Threads XHR</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

var worker = new SharedWorker("xhr_sharedworker.js");

worker.port.onmessage = function(event) {
  is(event.data, "done", "Got correct result");
  SimpleTest.finish();
}
worker.port.postMessage("worker_testXHR.txt");

SimpleTest.waitForExplicitFinish();

</script>
</body>
</html>
